pyproject.toml
[tool.poetry]
name = "raphael"
version = "0.0.1"
description = "Health misinformation from videos"
authors = ["Full Fact","Google"]
readme = "README.md"
license = "Proprietary"
packages = [
    { include = "health_misinfo_shared", from = "src" },
    { include = "raphael_backend_flask", from = "src" },
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
python-dotenv = "^1.0.0"
structlog = "^23.1.0"
pydantic = ">1.7"
google-cloud-aiplatform = "^1.38"
pandas = "1.5"

# For youtube access:
google-api-python-client="^2.118.0"
google-auth-oauthlib="^1.2.0"
google-auth-httplib2="^0.2.0"
langdetect="^1.0.9"
flask = "^3.0.3"
flask-httpauth = "^4.8.0"
flask-cors = "^4.0.0"
gunicorn = "^22.0.0"
fsspec = "^2024.5.0"
gcsfs = "^2024.5.0"
json-repair = "^0.28.4"
yt-dlp = "^2024.8.6"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
debugpy = "^1.6.7"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = ">0.9"
pre-commit = "^3.3.3"
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
scikit-learn = "^1.5.0"
rouge-score = "^0.1.2"

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]

[tool.isort]
profile = "black"

[tool.mypy]
warn_unused_configs = true
disallow_untyped_calls = true
check_untyped_defs = true
